Add number guessing game
The file guess.py contains a working solution for the
guessing game from section 6, lesson 21. In this version of
the game, the computer attempts to guess the number that the
player has in memory.

This exercise demonstrates several language techniques
such as looping with a `while` loop and defining functions.